refactor(json-crdt-patch): 💡 move CRDT encoder/decoder base clases util
