+++ src/json-crdt-patch/util/binary/CrdtEncoder.ts
